Bump to 0.0.5. Formulas are now used for all computation, including intial values, maximum values and all formula.
This is a fairly large reworking of the internals to support more
dynamic configuration everywhere. There is some major work to do
though in terms of removing rough edges on this though, for example
currently references in formulas for "initial" values are rejected
because we need to add validation for circular references before we
could safely allow that.

The other issue is that the interface for programmatic usage is pretty
bad now, which also makes testing a bit more annoying since you have to
pass around systems.models.Formula instances everywhere. The plan is to
add systems.simple module which will provide a cleaner interface for
programmatic users, although generally the expectation is that most folks
will continue to rely on the text format and parser instead of doing things
programmatically.